Removed some duplication in tests
